+++ src/index.js
@@ -1,4 +1,7 @@
 import { BASEBALL } from './constants.js';
+import { getRandom } from './functions/RandomNum.js';
+import { checkNum } from './functions/UserNumCheck.js';
+import { getBallCount, getStrikeCount, getResultText } from './functions/CountResult.js';
 
 export default function BaseballGame() {
   const submitBtn = document.getElementById('submit');
@@ -7,69 +10,7 @@ export default function BaseballGame() {
   const restartDiv = document.getElementById('restart');
   const restartBtn = document.getElementById('game-restart-button');
 
-  // 컴퓨터 랜덤숫자 반환
-  const getComputerNum = () => {
-    let numList = new Array();
-    let comNum = '';
-
-    while (numList.length < 3) {
-      const num = MissionUtils.Random.pickNumberInRange(1, 9);
-      if (!numList.includes(num)) {
-        numList.push(num);
-        comNum += String(num);
-      }
-    }
-
-    return comNum;
-  };
-
-  let computerInputNumbers = getComputerNum();
-
-  // check 1- 숫자인지 확인
-  const checktype = num => {
-    if (isNaN(Number(num))) {
-      alert(BASEBALL.ALERT_ONLY_NUM);
-      return true;
-    }
-  };
-
-  // check 2- 3자리인지 확인
-  const checkLen = num => {
-    if (num.length !== 3) {
-      alert(BASEBALL.ALERT_THREE_NUM);
-      return true;
-    }
-  };
-
-  // check3 - 숫자 범위 확인 (0 포함 확인)
-  const checkOnetoNine = num => {
-    if (num.includes(0)) {
-      alert(BASEBALL.ALERT_ONE_TO_NINE);
-      return true;
-    }
-  };
-
-  // check4 - 숫자 중복 여부 확인
-  const checkDuplication = num => {
-    if (num[0] === num[1] || num[1] === num[2] || num[0] === num[2]) {
-      alert(BASEBALL.ALERT_DUPLICATION);
-      return true;
-    }
-  };
-
-  // check All
-  const checkNum = num => {
-    if (
-      checktype(num) ||
-      checkLen(num) ||
-      checkOnetoNine(num) ||
-      checkDuplication(num)
-    ) {
-      return;
-    }
-
-    return true;
-  };
+  let computerInputNumbers = getRandom();
 
   // 사용자 수 입력받기
   const getUserNum = () => {
@@ -79,50 +20,6 @@ export default function BaseballGame() {
       return userInputNumbers;
     }
   };
-
-  // get Ball
-  const getBallCount = (computerNum, userNum) => {
-    let ball = 0;
-    let i;
-    for (i = 0; i < 3; i += 1) {
-      if (computerNum[i] !== userNum[i] && computerNum.includes(userNum[i])) {
-        ball += 1;
-      }
-    }
-
-    return ball;
-  };
-
-  // get Strike
-  const getStrikeCount = (computerNum, userNum) => {
-    let strike = 0;
-    let i;
-    for (i = 0; i < 3; i += 1) {
-      if (computerNum[i] === userNum[i]) {
-        strike += 1;
-      }
-    }
-
-    return strike;
-  };
-
-  // return ball, strike result
-  const getResultText = (strike, ball) => {
-    let result = '';
-    if (strike === 3) {
-      return BASEBALL.ANSWER;
-    }
-    if (strike === 0 && ball === 0) {
-      return BASEBALL.NOTHING;
-    }
-    if (ball !== 0) {
-      result += `${ball}${BASEBALL.BALL} `;
-    }
-    if (strike !== 0) {
-      result += `${strike}${BASEBALL.STRIKE}`;
-    }
-    return result;
-  }
 
   // play 함수
   this.play = (computerInputNumbers, userInputNumbers) => {
@@ -151,6 +48,7 @@ export default function BaseballGame() {
     let result = '';
 
     if (userNum !== undefined) {
+      console.log(computerInputNumbers);
       result = this.play(computerInputNumbers, userNum);
     }
     showResult(result);
@@ -164,7 +62,7 @@ export default function BaseballGame() {
     userInput.value = '';
     resultDiv.innerText = '';
 
-    computerInputNumbers = getComputerNum();
+    computerInputNumbers = getRandom();
   };
 
   submitBtn.addEventListener('click', startGame);
